# CoffeePeek.UserService/Controllers/UserController.cs
using CoffeePeek.Contract.Dtos.User;
using CoffeePeek.Contract.Requests.User;
using CoffeePeek.Contract.Response;
using CoffeePeek.Contract.Response.User;
using CoffeePeek.UserService.Extensions;
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace CoffeePeek.UserService.Controllers;

[ApiController]
[Route("api/[controller]")]
public class UserController(IMediator mediator) : Controller
{
    [HttpGet]
    [Authorize]
    [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status404NotFound)]
    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
    public Task<Response<UserDto>> GetProfile(CancellationToken cancellationToken)
    {
        var request = new GetProfileRequest(User.GetUserIdOrThrow());
        
        return mediator.Send(request, cancellationToken);
    }
    
    [HttpPut]
    [Authorize]
    [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
    public Task<Response<UpdateProfileResponse>> UpdateProfile([FromBody]UpdateProfileRequest request, CancellationToken cancellationToken)
    {
        var authenticatedRequest = request with { UserId = HttpContext.User.GetUserIdOrThrow() };
        return mediator.Send(authenticatedRequest, cancellationToken);
    }
    
    [HttpGet("Users")]
    [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
    public Task<Response<UserDto[]>> GetAllUsers(CancellationToken cancellationToken)
    {
        var request = new GetAllUsersRequest();
        return mediator.Send(request, cancellationToken);
    }

    [HttpDelete("{id:guid}")]
    [Authorize]
    [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
    public Task<Response<bool>> DeleteUser(Guid id, CancellationToken cancellationToken)
    {
        var request = new DeleteUserRequest(id);
        return mediator.Send(request, cancellationToken);
    }
}