# openapi.yaml
openapi: 3.0.3
info:
  title: CoffeePeek API
  description: API для платформы CoffeePeek - поиска и оценки кофеен
  version: 1.0.0
  contact:
    name: CoffeePeek Team
servers:
  - url: https://api.coffeepeek.com
    description: Production server
  - url: http://localhost:80
    description: Local development server
...
paths:
...
  /api/user:
    get:
      tags:
        - User
      summary: Получить профиль текущего пользователя
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - User
      summary: Обновить профиль пользователя
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/user/Users:
    get:
      tags:
        - User
      summary: Получить всех пользователей
      operationId: getAllUsers
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/user/{id}:
    delete:
      tags:
        - User
      summary: Удалить пользователя
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID пользователя
      responses:
        '200':
          description: Пользователь успешно удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== CoffeeShop Endpoints ====================
...
  /api/reviewcoffee/user/{id}:
    get:
      tags:
        - Review
      summary: Получить все отзывы пользователя
      operationId: getReviewsByUserId
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID пользователя
      responses:
        '200':
          description: Список отзывов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetReviewsByUserIdResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Moderation Endpoints ====================
  /api/moderation:
    get:
      tags:
        - Moderation
      summary: Получить все кофейни на модерации для текущего пользователя
      operationId: getCoffeeShopsInModerationByUserId
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список кофеен на модерации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoffeeShopsInModerationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Moderation
      summary: Отправить кофейню на модерацию
      operationId: sendCoffeeShopToModeration
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCoffeeShopToModerationRequest'
      responses:
        '200':
          description: Кофейня отправлена на модерацию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendCoffeeShopToModerationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Moderation
      summary: Обновить кофейню на модерации
      operationId: updateModerationCoffeeShop
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateModerationCoffeeShopRequest'
      responses:
        '200':
          description: Кофейня на модерации обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateModerationCoffeeShopResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/moderation/all:
    get:
      tags:
        - Moderation
      summary: Получить все кофейни на модерации (только для администраторов)
      operationId: getAllModerationShops
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список всех кофеен на модерации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoffeeShopsInModerationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (требуются права администратора)
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/moderation/status:
    put:
      tags:
        - Moderation
      summary: Обновить статус модерации кофейни (только для администраторов)
      operationId: updateModerationCoffeeShopStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
          description: ID кофейни на модерации
        - name: status
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ModerationStatus'
          description: Новый статус модерации
      responses:
        '200':
          description: Статус модерации обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (требуются права администратора)
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Internal Endpoints ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа
  schemas:
...
    BaseResponse:
      type: object
      properties:
        isSuccess:
          type: boolean
        message:
          type: string
        errorCode:
          type: string
          nullable: true
        traceId:
          type: string
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
    BooleanResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: boolean

    # Auth Schemas
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        userName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        token:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
          nullable: true
        about:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        photoUrl:
          type: string
          nullable: true
        reviewCount:
          type: integer
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserDto'
    UsersListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/UserDto'
    UpdateProfileRequest:
      type: object
      properties:
        userName:
          type: string
        about:
          type: string
          nullable: true
    UpdateProfileResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object

    # CoffeeShop Schemas
...
    ShopContactDto:
      type: object
      properties:
        phone:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        instagram:
          type: string
          nullable: true
    ScheduleDto:
      type: object
      properties:
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        intervals:
          type: array
          items:
            $ref: '#/components/schemas/ShopScheduleIntervalDto'
    ShopScheduleIntervalDto:
      type: object
      properties:
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
...
    CoffeeShopReviewDto:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        userId:
          type: integer
        header:
          type: string
        comment:
          type: string
        ratingCoffee:
          type: number
          format: decimal
        ratingService:
          type: number
          format: decimal
        ratingPlace:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        shopName:
          type: string
...
    GetReviewsByUserIdResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                reviewDtos:
                  type: array
                  items:
                    $ref: '#/components/schemas/CoffeeShopReviewDto'

    # Moderation Schemas
    ModerationStatus:
      type: string
      enum: [Pending, Approved]
    ModerationShopDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        notValidatedAddress:
          type: string
        address:
          type: integer
          nullable: true
        moderationStatus:
          $ref: '#/components/schemas/ModerationStatus'
        status:
          type: string
        shopContact:
          $ref: '#/components/schemas/ShopContactDto'
          nullable: true
        shopPhotos:
          type: array
          items:
            type: string
            format: uri
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDto'
    SendCoffeeShopToModerationRequest:
      type: object
      required:
        - name
        - notValidatedAddress
      properties:
        name:
          type: string
        notValidatedAddress:
          type: string
        shopContact:
          $ref: '#/components/schemas/ShopContactDto'
        shopPhotos:
          type: array
          items:
            type: string
            format: uri
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDto'
    UpdateModerationCoffeeShopRequest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        notValidatedAddress:
          type: string
        shopContact:
          $ref: '#/components/schemas/ShopContactDto'
        shopPhotos:
          type: array
          items:
            type: string
            format: uri
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDto'
    SendCoffeeShopToModerationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
    UpdateModerationCoffeeShopResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
    GetCoffeeShopsInModerationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                moderationShop:
                  type: array
                  items:
                    $ref: '#/components/schemas/ModerationShopDto'

    # Internal Schemas
...
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
...